refactor: typeorm repo 이름 변경
