+++ src/modules/posts/posts.module.ts
@@ -2,7 +2,7 @@ import { Logger, Module } from '@nestjs/common';
 import { PostsService } from './posts.service';
 import { PostsController } from './posts.controller';
 import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module';
-import { PostsRepository } from './posts.repository';
+import { PostsRepository } from './posts-typeorm.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../users/entities/user.entity';
 import { Category } from '../categorys/entities/category.entity';
