+++ src/test/repository/comments.repository.spec.ts
@@ -5,7 +5,7 @@ import { UsersRepository } from '../../modules/users/users-typeorm.repository';
 import { TestTypeOrmModule } from '../db/test-db.module';
 import { User } from '../../modules/users/entities/user.entity';
 import { Comment } from '../../modules/comments/entities/comment.entity';
-import { PostsRepository } from '../../modules/posts/posts.repository';
+import { PostsRepository } from '../../modules/posts/posts-typeorm.repository';
 import { Post } from '../../modules/posts/entities/post.entity';
 
 describe('CommentsRepository', () => {
